ref move app component
